Use the last set position for clicking on OSX
There is an issue with setting the position then clicking on OSX.
Because events are not posted synchronously, it will take some time
until the mouse move event is processed. When the mouse buttons
are pressed/released before the event is processed, the current
mouse position is read to determine where to click, but that will
read the old position before the mouse move occured.

This is a bit of an ugly workaround and a better fix is needed.